requirements.txt
gunicorn==20.0.4
Flask==2.1.0
farm-haystack==1.12.2